♻️ Move types into invidual files
