+++ packages/inference/src/tasks/audio/audioClassification.ts
@@ -1,5 +1,26 @@
-import type { AudioClassificationArgs, AudioClassificationReturn, Options } from "../../types";
+import type { BaseArgs, Options } from "../../types";
 import { request } from "../custom/request";
+
+export type AudioClassificationArgs = BaseArgs & {
+	/**
+	 * Binary audio data
+	 */
+	data: Blob | ArrayBuffer;
+};
+
+export interface AudioClassificationReturnValue {
+	/**
+	 * The label for the class (model specific)
+	 */
+	label: string;
+
+	/**
+	 * A float that represents how likely it is that the audio file belongs to this class.
+	 */
+	score: number;
+}
+
+export type AudioClassificationReturn = AudioClassificationReturnValue[];
 
 /**
  * This task reads some audio input and outputs the likelihood of classes.
