+++ packages/inference/src/tasks/cv/imageSegmentation.ts
@@ -1,5 +1,29 @@
-import type { ImageSegmentationArgs, ImageSegmentationReturn, Options } from "../../types";
+import type { BaseArgs, Options } from "../../types";
 import { request } from "../custom/request";
+
+export type ImageSegmentationArgs = BaseArgs & {
+	/**
+	 * Binary image data
+	 */
+	data: Blob | ArrayBuffer;
+};
+
+export interface ImageSegmentationReturnValue {
+	/**
+	 * The label for the class (model specific) of a segment.
+	 */
+	label: string;
+	/**
+	 * A str (base64 str of a single channel black-and-white img) representing the mask of a segment.
+	 */
+	mask: string;
+	/**
+	 * A float that represents how likely it is that the detected object belongs to the given class.
+	 */
+	score: number;
+}
+
+export type ImageSegmentationReturn = ImageSegmentationReturnValue[];
 
 /**
  * This task reads some image input and outputs the likelihood of classes & bounding boxes of detected objects.
