+++ packages/inference/src/tasks/cv/imageToText.ts
@@ -1,5 +1,19 @@
-import type { ImageToTextArgs, ImageToTextReturn, Options } from "../../types";
+import type { BaseArgs, Options } from "../../types";
 import { request } from "../custom/request";
+
+export type ImageToTextArgs = BaseArgs & {
+	/**
+	 * Binary image data
+	 */
+	data: Blob | ArrayBuffer;
+};
+
+export interface ImageToTextReturn {
+	/**
+	 * The generated caption
+	 */
+	generated_text: string;
+}
 
 /**
  * This task reads some image input and outputs the text caption.
