+++ packages/inference/src/tasks/cv/objectDetection.ts
@@ -1,5 +1,35 @@
 import { request } from "../custom/request";
-import type { ObjectDetectionArgs, ObjectDetectionReturn, Options } from "../../types";
+import type { BaseArgs, Options } from "../../types";
+
+export type ObjectDetectionArgs = BaseArgs & {
+	/**
+	 * Binary image data
+	 */
+	data: Blob | ArrayBuffer;
+};
+
+export interface ObjectDetectionReturnValue {
+	/**
+	 * A dict (with keys [xmin,ymin,xmax,ymax]) representing the bounding box of a detected object.
+	 */
+	box: {
+		xmax: number;
+		xmin: number;
+		ymax: number;
+		ymin: number;
+	};
+	/**
+	 * The label for the class (model specific) of a detected object.
+	 */
+	label: string;
+
+	/**
+	 * A float that represents how likely it is that the detected object belongs to the given class.
+	 */
+	score: number;
+}
+
+export type ObjectDetectionReturn = ObjectDetectionReturnValue[];
 
 /**
  * This task reads some image input and outputs the likelihood of classes & bounding boxes of detected objects.
