+++ packages/inference/src/tasks/nlp/textClassification.ts
@@ -1,5 +1,23 @@
-import type { Options, TextClassificationArgs, TextClassificationReturn } from "../../types";
+import type { BaseArgs, Options } from "../../types";
 import { request } from "../custom/request";
+
+export type TextClassificationArgs = BaseArgs & {
+	/**
+	 * A string to be classified
+	 */
+	inputs: string;
+};
+
+export type TextClassificationReturn = {
+	/**
+	 * The label for the class (model specific)
+	 */
+	label: string;
+	/**
+	 * A floats that represents how likely is that the text belongs to this class.
+	 */
+	score: number;
+}[];
 
 /**
  * Usually used for sentiment-analysis this will output the likelihood of classes of an input. Recommended model: distilbert-base-uncased-finetuned-sst-2-english
