+++ packages/inference/src/tasks/nlp/tokenClassification.ts
@@ -1,11 +1,54 @@
-import type {
-	Options,
-	TokenClassificationArgs,
-	TokenClassificationReturn,
-	TokenClassificationReturnValue,
-} from "../../types";
+import type { BaseArgs, Options } from "../../types";
 import { toArray } from "../../utils/toArray";
 import { request } from "../custom/request";
+
+export type TokenClassificationArgs = BaseArgs & {
+	/**
+	 * A string to be classified
+	 */
+	inputs: string;
+	parameters?: {
+		/**
+		 * (Default: simple). There are several aggregation strategies:
+		 *
+		 * none: Every token gets classified without further aggregation.
+		 *
+		 * simple: Entities are grouped according to the default schema (B-, I- tags get merged when the tag is similar).
+		 *
+		 * first: Same as the simple strategy except words cannot end up with different tags. Words will use the tag of the first token when there is ambiguity.
+		 *
+		 * average: Same as the simple strategy except words cannot end up with different tags. Scores are averaged across tokens and then the maximum label is applied.
+		 *
+		 * max: Same as the simple strategy except words cannot end up with different tags. Word entity will be the token with the maximum score.
+		 */
+		aggregation_strategy?: "none" | "simple" | "first" | "average" | "max";
+	};
+};
+
+export interface TokenClassificationReturnValue {
+	/**
+	 * The offset stringwise where the answer is located. Useful to disambiguate if word occurs multiple times.
+	 */
+	end: number;
+	/**
+	 * The type for the entity being recognized (model specific).
+	 */
+	entity_group: string;
+	/**
+	 * How likely the entity was recognized.
+	 */
+	score: number;
+	/**
+	 * The offset stringwise where the answer is located. Useful to disambiguate if word occurs multiple times.
+	 */
+	start: number;
+	/**
+	 * The string that was captured
+	 */
+	word: string;
+}
+
+export type TokenClassificationReturn = TokenClassificationReturnValue[];
 
 /**
  * Usually used for sentence parsing, either grammatical, or Named Entity Recognition (NER) to understand keywords contained within text. Recommended model: dbmdz/bert-large-cased-finetuned-conll03-english
@@ -14,7 +57,7 @@ export async function tokenClassification(
 	args: TokenClassificationArgs,
 	options?: Options
 ): Promise<TokenClassificationReturn> {
-	const res = toArray(await request<TokenClassificationReturnValue | TokenClassificationReturn>(args, options));
+	const res = toArray(await request<TokenClassificationReturn[number] | TokenClassificationReturn>(args, options));
 	const isValidOutput =
 		Array.isArray(res) &&
 		res.every(
